test(api): add SessionService specs
